Add accum functionality to mxm
The matrix multiply function was changed to try and catch with the
accum operator. This should later be updated so that no try/catch
is needed for accum.

The difficuty arises in the use of get_value_at in the sparse
matrix object. A value cannot be returned when the index isn't
valid.